+++ 송창석/ebay/src/routes/Join/UserJoinContainer.jsx
@@ -1,24 +1,12 @@
-import React from 'react';
-import useFormStore from '../../store/useFormStore';
+import React, { useContext } from 'react';
+import { UserJoinContext } from '../../App';
 import UserJoinPresenter from './UserJoinPresenter';
 
-export const UserJoinContext = React.createContext(null)
-
 const UserJoinContainer = () => {
-  const UserJoinValues = useFormStore({
-    name: "",
-    email: "",
-    password: "",
-    gender: "",
-    year: "",
-    month: "",
-    day: ""
-  })
+  const provide = useContext(UserJoinContext)
 
   return (
-    <UserJoinContext.Provider value={UserJoinValues}>
-      <UserJoinPresenter/>
-    </UserJoinContext.Provider>
+    <UserJoinPresenter state={provide}/>
   );
 };
